tydyng up inout module
